try workaround for windows
This commit adds ggml.c to the test-backend-ops compilation unit as on
windows the linker cannot resolve some of the functions that have been
moved into test-backend-ops, like ggml_hash_set_new. I'm not sure we
will keep this moving but for now this will at least allow the test to
be run on windows as well.