sampling : add support for GPU sampling (wip) [no ci]
This commit adds support for GPU sampling in llama.cpp.

The motivation for this feature is to enable sampling on the GPU
and takes advantage of the fact that the logits are already on the GPU
enabling these GPU samplers to efficiently operate on the logits
without needing to first copy them from device memory.

It is possible to have a GPU sampler to also select a token in which
case there is a function named llama_get_sampled_token(ctx) which will
retrieve it. Not all GPU sampler will select/sample a token, some might
just want to perform temperature scaling and/or top-k filtering and then
perform the remaining sampling on the CPU. In this case only the GPU
samplers are added to the context parameters sampler chain, and the CPU
sampler are configured just like they are normally.

Testing:
Download a model for testing:
```console
$ cd models && wget https://huggingface.co/ggml-org/models/resolve/main/tinyllamas/stories15M-q4_0.gguf
```
Building the test:
```console
$ cmake --build build-gpu-sampling --target test-gpu-sampling -j8
```
Runing the test:
```
$ env LLAMACPP_TEST_MODELFILE=../models/stories15M-q4_0.gguf \
    ctest --test-dir build-gpu-sampling -R 'test-gpu-sampling' -V
```